chore(weave): extract clickhouse/tables.py with CHInfraProtocol
Move table create, update, query, stream, and stats methods into TablesMixin.
